DataReader is now ReadableData; add new DataReader
